image resizer

+++ src/views/TrainPane.java
@@ -119,7 +119,7 @@ public class TrainPane extends JPanel
 		panel2.add(label3);
 		
 		employedSpinner = new JSpinner();
-		employedSpinner.setModel(new SpinnerNumberModel(new Integer(20), 1, null, new Integer(5)));
+		employedSpinner.setModel(new SpinnerNumberModel(new Integer(10), 1, null, new Integer(5)));
 		employedSpinner.setBounds(332, 163, 67, 30);
 		employedSpinner.setBorder(null);
 		employedSpinner.setFont(new Font("Century Gothic", Font.PLAIN, 15));
@@ -155,7 +155,7 @@ public class TrainPane extends JPanel
 		panel4.add(label5);
 
 		onlookerSpinner = new JSpinner();
-		onlookerSpinner.setModel(new SpinnerNumberModel(new Integer(30), 1, null, new Integer(5)));
+		onlookerSpinner.setModel(new SpinnerNumberModel(new Integer(15), 1, null, new Integer(5)));
 		onlookerSpinner.setBounds(332, 219, 67, 30);
 		onlookerSpinner.setBorder(null);
 		onlookerSpinner.setFont(new Font("Century Gothic", Font.PLAIN, 15));
